Moved string casing functions to StringExtensions.Casing.cs
