+++ Cave.Extensions/StringExtensions.cs
@@ -709,25 +709,6 @@ public static partial class StringExtensions
         return result.ToString();
     }
 
-    /// <summary>Builds a camel case name split at invalid characters and upper case letters.</summary>
-    /// <param name="validChars">Valid characters.</param>
-    /// <param name="splitter">Character used to split parts.</param>
-    /// <param name="text">The text to use.</param>
-    /// <returns>A camel case version of text.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string GetCamelCaseName(this string text, string validChars, char splitter)
-    {
-        text = text.ReplaceInvalidChars(validChars, $"{splitter}");
-        var parts = text.Split(splitter).SelectMany(s => s.SplitCasing());
-        return parts.ToArray().JoinCamelCase();
-    }
-
-    /// <summary>Builds a camel case name split at invalid characters and upper case letters.</summary>
-    /// <param name="text">The text to use.</param>
-    /// <returns>A camel case version of text.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string GetCamelCaseName(this string text) => GetCamelCaseName(text, ASCII.Strings.Letters + ASCII.Strings.Digits, '_');
-
     /// <summary>Retrieves only invalidated chars from a string.</summary>
     /// <param name="text">The text.</param>
     /// <param name="validChars">The string with the valid chars.</param>
@@ -751,64 +732,6 @@ public static partial class StringExtensions
 
         return result.ToString();
     }
-
-    /// <summary>Builds a kebab case name split at invalid characters and upper case letters.</summary>
-    /// <param name="validChars">Valid characters.</param>
-    /// <param name="splitter">Character used to split parts.</param>
-    /// <param name="text">The text to use.</param>
-    /// <returns>A kebab case version of text.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string GetKebabCaseName(this string text, string validChars, char splitter)
-    {
-        text = text.ReplaceInvalidChars(validChars, $"{splitter}");
-        var parts = text.Split(splitter).SelectMany(s => s.SplitCasing());
-        return parts.ToArray().JoinKebabCase();
-    }
-
-    /// <summary>Builds a camel case name split at invalid characters and upper case letters.</summary>
-    /// <param name="text">The text to use.</param>
-    /// <returns>A camel case version of text.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string GetKebabCaseName(this string text) => GetKebabCaseName(text, ASCII.Strings.Letters + ASCII.Strings.Digits, '-');
-
-    /// <summary>Builds a camel case name split at invalid characters and upper case letters.</summary>
-    /// <param name="validChars">Valid characters.</param>
-    /// <param name="splitter">Character used to split parts.</param>
-    /// <param name="text">The text to use.</param>
-    /// <returns>A camel case version of text.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string GetLowerCamelCaseName(this string text, string validChars, char splitter)
-    {
-        text = text.ReplaceInvalidChars(validChars, $"{splitter}");
-        var parts = text.Split(splitter).SelectMany(s => s.SplitCasing());
-        return parts.ToArray().JoinLowerCamelCase();
-    }
-
-    /// <summary>Builds a camel case name split at invalid characters and upper case letters.</summary>
-    /// <param name="text">The text to use.</param>
-    /// <returns>A camel case version of text.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string GetLowerCamelCaseName(this string text) => GetLowerCamelCaseName(text, ASCII.Strings.Letters + ASCII.Strings.Digits, '_');
-
-    /// <summary>Builds a camel case name split at invalid characters and upper case letters.</summary>
-    /// <param name="validChars">Valid characters.</param>
-    /// <param name="splitter">Character used to split parts.</param>
-    /// <param name="text">The text to use.</param>
-    /// <returns>A camel case version of text.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string GetSnakeCaseName(this string text, string validChars, char splitter)
-    {
-        text ??= string.Empty;
-        text = text.ReplaceInvalidChars(validChars, $"{splitter}");
-        var parts = text.Split(splitter).SelectMany(s => s.SplitCasing());
-        return parts.ToArray().JoinSnakeCase();
-    }
-
-    /// <summary>Builds a camel case name split at invalid characters and upper case letters.</summary>
-    /// <param name="text">The text to use.</param>
-    /// <returns>A camel case version of text.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string GetSnakeCaseName(this string text) => GetSnakeCaseName(text, ASCII.Strings.Letters + ASCII.Strings.Digits, '_');
 
     /// <summary>Gets a part of a string.</summary>
     /// <param name="data">Data to parse.</param>
@@ -1205,220 +1128,6 @@ public static partial class StringExtensions
         return result.ToString();
     }
 
-    /// <summary>Joins the strings with camel casing.</summary>
-    /// <param name="parts">The parts.</param>
-    /// <param name="culture">The culture info.</param>
-    /// <returns>The joined string.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string JoinCamelCase(this string[] parts, CultureInfo culture = null)
-    {
-        if ((parts == null) || (parts.Length == 0))
-        {
-            return string.Empty;
-        }
-
-        culture ??= CultureInfo.CurrentCulture;
-
-        var result = new StringBuilder();
-        foreach (var part in parts)
-        {
-            var t = part.Trim();
-            if (t.Length < 1)
-            {
-                continue;
-            }
-
-            _ = result.Append(t[0].ToUpper(culture));
-            if (t.Length > 1)
-            {
-                _ = result.Append(t[1..].ToLower(culture));
-            }
-        }
-
-        return result.ToString();
-    }
-
-    /// <summary>Joins the strings with camel casing.</summary>
-    /// <param name="parts">The parts.</param>
-    /// <param name="culture">The culture info.</param>
-    /// <returns>The joned string.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string JoinCamelCase(this IEnumerable parts, CultureInfo culture = null)
-    {
-        if (parts == null)
-        {
-            return string.Empty;
-        }
-
-        culture ??= CultureInfo.CurrentCulture;
-
-        var result = new StringBuilder();
-        foreach (var part in parts)
-        {
-            var t = part is IFormattable formattable ? formattable.ToString(null, culture) : $"{part}";
-            if (t.Length < 1)
-            {
-                continue;
-            }
-
-            _ = result.Append(t[0].ToUpper(culture));
-            if (t.Length > 1)
-            {
-                _ = result.Append(t[1..].ToLower(culture));
-            }
-        }
-
-        return result.ToString();
-    }
-
-    /// <summary>Joins the strings using kebab case.</summary>
-    /// <param name="parts">The parts.</param>
-    /// <param name="culture">The culture info.</param>
-    /// <returns>The joned string.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string JoinKebabCase(this string[] parts, CultureInfo culture = null)
-    {
-        if ((parts == null) || (parts.Length == 0))
-        {
-            return string.Empty;
-        }
-
-        culture ??= CultureInfo.CurrentCulture;
-
-        var result = new StringBuilder();
-        foreach (var part in parts)
-        {
-            var t = part.Trim();
-            if (t.Length < 1)
-            {
-                continue;
-            }
-
-            if (result.Length > 0)
-            {
-                _ = result.Append('-');
-            }
-
-            _ = result.Append(t.ToLower(culture));
-        }
-
-        return result.ToString();
-    }
-
-    /// <summary>Joins the strings using kebab case.</summary>
-    /// <param name="parts">The parts.</param>
-    /// <param name="culture">The culture info.</param>
-    /// <returns>The joned string.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string JoinKebabCase(this IEnumerable parts, CultureInfo culture = null)
-    {
-        if (parts == null)
-        {
-            return string.Empty;
-        }
-
-        culture ??= CultureInfo.CurrentCulture;
-
-        var result = new StringBuilder();
-        foreach (var part in parts)
-        {
-            var t = part is IFormattable formattable ? formattable.ToString(null, culture) : $"{part}";
-            if (t.Length < 1)
-            {
-                continue;
-            }
-
-            if (result.Length > 0)
-            {
-                _ = result.Append('_');
-            }
-
-            _ = result.Append(t.ToLower(culture));
-        }
-
-        return result.ToString();
-    }
-
-    /// <summary>Joins the strings with camel casing.</summary>
-    /// <param name="parts">The parts.</param>
-    /// <param name="culture">The culture info.</param>
-    /// <returns>The joned string.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string JoinLowerCamelCase(this string[] parts, CultureInfo culture = null)
-    {
-        if ((parts == null) || (parts.Length == 0))
-        {
-            return string.Empty;
-        }
-
-        culture ??= CultureInfo.CurrentCulture;
-
-        var result = new StringBuilder();
-        foreach (var part in parts)
-        {
-            var t = part.Trim();
-            if (t.Length < 1)
-            {
-                continue;
-            }
-
-            if (result.Length > 0)
-            {
-                _ = result.Append(t[0].ToUpper(culture));
-                if (t.Length > 1)
-                {
-                    _ = result.Append(t[1..].ToLower(culture));
-                }
-            }
-            else
-            {
-                _ = result.Append(t.ToLower(culture));
-            }
-        }
-
-        return result.ToString();
-    }
-
-    /// <summary>Joins the strings with camel casing.</summary>
-    /// <param name="parts">The parts.</param>
-    /// <param name="culture">The culture info.</param>
-    /// <returns>The joned string.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string JoinLowerCamelCase(this IEnumerable parts, CultureInfo culture = null)
-    {
-        if (parts == null)
-        {
-            return string.Empty;
-        }
-
-        culture ??= CultureInfo.CurrentCulture;
-
-        var result = new StringBuilder();
-        foreach (var part in parts)
-        {
-            var t = part is IFormattable formattable ? formattable.ToString(null, culture) : $"{part}";
-            if (t.Length < 1)
-            {
-                continue;
-            }
-
-            if (result.Length > 0)
-            {
-                _ = result.Append(t[0].ToUpper(culture));
-                if (t.Length > 1)
-                {
-                    _ = result.Append(t[1..].ToLower(culture));
-                }
-            }
-            else
-            {
-                _ = result.Append(t.ToLower(culture));
-            }
-        }
-
-        return result.ToString();
-    }
-
     /// <summary>Joins a collection to a string with newlines for all systems.</summary>
     /// <param name="texts">The string collection.</param>
     /// <returns>Returns a new string.</returns>
@@ -1430,74 +1139,6 @@ public static partial class StringExtensions
     /// <returns>Returns a new string.</returns>
     [MethodImpl((MethodImplOptions)256)]
     public static string JoinNewLine(this IEnumerable array) => Join(array, "\r\n", null);
-
-    /// <summary>Joins the strings using snake case.</summary>
-    /// <param name="parts">The parts.</param>
-    /// <param name="culture">The culture info.</param>
-    /// <returns>The joned string.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string JoinSnakeCase(this string[] parts, CultureInfo culture = null)
-    {
-        if ((parts == null) || (parts.Length == 0))
-        {
-            return string.Empty;
-        }
-
-        culture ??= CultureInfo.CurrentCulture;
-
-        var result = new StringBuilder();
-        foreach (var part in parts)
-        {
-            var t = part.Trim();
-            if (t.Length < 1)
-            {
-                continue;
-            }
-
-            if (result.Length > 0)
-            {
-                _ = result.Append('_');
-            }
-
-            _ = result.Append(t.ToLower(culture));
-        }
-
-        return result.ToString();
-    }
-
-    /// <summary>Joins the strings using snake case.</summary>
-    /// <param name="parts">The parts.</param>
-    /// <param name="culture">The culture info.</param>
-    /// <returns>The joned string.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string JoinSnakeCase(this IEnumerable parts, CultureInfo culture = null)
-    {
-        if (parts == null)
-        {
-            return string.Empty;
-        }
-
-        culture ??= CultureInfo.CurrentCulture;
-
-        var result = new StringBuilder();
-        foreach (var part in parts)
-        {
-            var t = part is IFormattable formattable ? formattable.ToString(null, culture) : $"{part}";
-            if (t.Length < 1)
-            {
-                continue;
-            }
-
-            if (result.Length > 0)
-            {
-                _ = result.Append('_');
-            }
-
-            _ = result.Append(t.ToLower(culture));
-        }
-
-        return result.ToString();
-    }
 
     /// <summary>Parses a binary size string created by <see cref="FormatSize(double, IFormatProvider)"/> or <see cref="FormatBinarySize(double, IFormatProvider)"/>.</summary>
     /// <param name="value">The value string.</param>
@@ -2165,45 +1806,6 @@ public static partial class StringExtensions
     /// <returns>The string array.</returns>
     [MethodImpl((MethodImplOptions)256)]
     public static string[] SplitAt(this string text, params int[] indices) => SplitAt(text, (IEnumerable<int>)indices);
-
-    /// <summary>Splits a string at character casing.</summary>
-    /// <param name="text">The text.</param>
-    /// <returns>The string array.</returns>
-    [Obsolete("Use SplitCasing() instead!")]
-    public static string[] SplitCamelCase(this string text) => SplitCasing(text);
-
-    /// <summary>Splits a string at character casing.</summary>
-    /// <param name="text">The text.</param>
-    /// <returns>The string array.</returns>
-    [MethodImpl((MethodImplOptions)256)]
-    public static string[] SplitCasing(this string text)
-    {
-        if (text == null)
-        {
-            return ArrayExtension.Empty<string>();
-        }
-
-        var splits = new List<int>();
-        var casing = Case.Upper;
-        for (var current = 1; current < text.Length; current++)
-        {
-            var lastCasing = casing;
-            var c = text[current];
-            casing = char.IsUpper(c) ? Case.Upper : char.IsDigit(c) ? Case.Digit : Case.Default;
-            if (casing != lastCasing)
-            {
-                if (lastCasing == Case.Digit || casing != Case.Default)
-                {
-                    if (current > 1)
-                    {
-                        splits.Add(current);
-                    }
-                }
-            }
-        }
-
-        return SplitAt(text, splits);
-    }
 
     /// <summary>Splits a string at the specified separators and allows to keep the separators in the list.</summary>
     /// <param name="text">The text.</param>
